requirements.txt
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
sentence-transformers ##opensource embedding model to generate vector embedding
## sentence-transformers==2.2.2 model uses HF hub to download the model
langchain-pinecone
langchain-community
langchain-experimental
langchain-google-genai
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
-e .
